+++ Clase2/Expresiones.java
@@ -1,9 +1,9 @@
 public class Expresiones {
 
     public static void main(String[] args){
-   int A = 4;
-   int B = 5;
-   int C = 1;
+   double A = 4;
+   double B = 5;
+   double C = 1;
     System.out.println((((B + C) /2 * A +10) *3 * B) -6);
     }
 }
